protozero: support decoding fields up to 2**24 - 1
Allow parsing of larger field ids. This is necessary after
supporting statsd protos in the trace, which use high field
ids for OEM atoms.

This change keeps the protozero::Field to 16 bytes total and
uses the existing padding space to extend the id_ field to 24 bits.

Benchmark:
----------
On an Intel MBP 16 there is no measureable difference.
Actually in some cases this patch seems to speed up
marginally the benchmarks.

```
$ out/mac_release/perfetto_benchmarks.new --benchmark_filter=BM_Protozero.* --benchmark_repetitions=10

Before this patch:
------------------
BM_Protozero_Simple_Libprotobuf_median         240 ns          239 ns           10
BM_Protozero_Simple_Protozero_median           253 ns          253 ns           10
BM_Protozero_Simple_SpeedOfLight_median       57.6 ns         57.5 ns           10
BM_Protozero_Nested_Libprotobuf_median        1146 ns         1143 ns           10
BM_Protozero_Nested_Protozero_median           581 ns          580 ns           10
BM_Protozero_Nested_SpeedOfLight_median       76.5 ns         76.4 ns           10
BM_ProtozeroMessageFilter_median         130139257 ns    129474100 ns           10 bytes_per_second=196.615M/s

After this patch:
-----------------
BM_Protozero_Simple_Libprotobuf_median         224 ns          223 ns           10
BM_Protozero_Simple_Protozero_median           255 ns          254 ns           10
BM_Protozero_Simple_SpeedOfLight_median       54.1 ns         53.9 ns
BM_Protozero_Nested_Libprotobuf_median        1153 ns         1147 ns           10
BM_Protozero_Nested_Protozero_median           561 ns          559 ns           10
BM_Protozero_Nested_SpeedOfLight_median       74.1 ns         74.0 ns           10
BM_ProtozeroMessageFilter_median         124684818 ns    124307100 ns           10 bytes_per_second=204.786M/s
```

Bug: https://github.com/google/perfetto/issues/510
Test: perfetto_unittests --gtest_filter=ProtoDecoderTest.SkipBigFieldIds

+++ src/trace_processor/importers/proto/track_event_parser.h
@@ -129,7 +129,7 @@ class TrackEventParser {
   ChromeStringLookup chrome_string_lookup_;
   std::array<StringId, 4> counter_unit_ids_;
 
-  std::vector<uint16_t> reflect_fields_;
+  std::vector<uint32_t> reflect_fields_;
 
   ActiveChromeProcessesTracker active_chrome_processes_tracker_;
 };
